refactor(explorer): move actions to user dialog
