+++ packages/dm-core-plugins/src/explorer/components/dialogs/DeleteDialog.tsx
@@ -1,12 +1,18 @@
-import { Dialog, TreeNode, useDMSS } from '@development-framework/dm-core'
+import {
+  Dialog,
+  ErrorResponse,
+  TreeNode,
+  useDMSS,
+} from '@development-framework/dm-core'
 import { Button, Progress } from '@equinor/eds-core-react'
+import { AxiosError } from 'axios'
 import React from 'react'
+import { toast } from 'react-toastify'
 import { EDialog } from '../../types'
 import {
   STANDARD_DIALOG_HEIGHT,
   STANDARD_DIALOG_WIDTH,
 } from '../context-menu/NodeRightClickMenu'
-import { DeleteAction } from '../context-menu/utils/contextMenuActions'
 
 type TProps = {
   setDialogId: (id: EDialog | undefined) => void
@@ -18,6 +24,27 @@ type TProps = {
 const DeleteDialog = (props: TProps) => {
   const { setDialogId, loading, setLoading, node } = props
   const dmssAPI = useDMSS()
+
+  const handleDelete = async () => {
+    setLoading(true)
+    await dmssAPI
+      .documentRemove({
+        address: `${node.dataSource}/${node.pathFromRootPackage()}`,
+      })
+      .then(() => {
+        node.remove()
+        toast.success('Deleted')
+      })
+      .catch((error: AxiosError<ErrorResponse>) => {
+        console.error(error)
+        toast.error('Failed to delete')
+      })
+      .finally(() => {
+        setLoading(false)
+        setDialogId(undefined)
+      })
+  }
+
   return (
     <Dialog
       open={true}
@@ -34,14 +61,7 @@ const DeleteDialog = (props: TProps) => {
         <b>{node.type}</b>?
       </Dialog.CustomContent>
       <Dialog.Actions>
-        <Button
-          color="danger"
-          onClick={async () => {
-            await DeleteAction(node, dmssAPI, setLoading)
-            await node.remove()
-            setDialogId(undefined)
-          }}
-        >
+        <Button color="danger" onClick={handleDelete}>
           {loading ? <Progress.Dots /> : 'Delete'}
         </Button>
         <Button variant="outlined" onClick={() => setDialogId(undefined)}>
